+++ resources/views/partials/footer.blade.php
@@ -1,6 +1,6 @@
 @php( $text_arabe = 'ماع سيراب ةرزجم 1961' )
 <span id="website-footer">
-    <a class="footer-text couvrefeu-text" style="position:fixed; bottom: 120px; left:-100px; font-size:28px;" href="{{ home_url('/') }}">{{ $text_arabe }}</a>
+    <a class="footer-text couvrefeu-text" style="position:fixed; bottom: 120px; left:-100px; font-size:28px; color:black;" href="{{ home_url('/') }}">{{ $text_arabe }}</a>
 </span>
 
 <footer class="content-info navbar fixed-bottom" style="padding:0">
@@ -9,3 +9,21 @@
         @php(dynamic_sidebar('sidebar-footer'))
     </div>
 </footer>
+<script type="text/javascript">
+function scroll_style() {
+    var window_top = jQuery(window).scrollTop();
+    var div_top = jQuery('.year-block.bg-warning').offset().top;
+    console.log(window_top);
+    console.log(div_top);
+    if (window_top > div_top-400){
+        jQuery('#website-footer a.footer-text').css({"color":"white"});
+    }else{
+        jQuery('#website-footer a.footer-text').css({"color":"black"});
+    }
+}
+jQuery(function() {
+    jQuery(window).scroll(scroll_style);
+    scroll_style();
+});
+
+</script>